removed warnings in assertions

+++ src/disp/display1.cc
@@ -184,7 +184,7 @@ void Display1::drawSinglesF(const SingleSet & set) const
                 ++ptr;
             }
         }
-        assert_true( ptr-flu <= cnt );
+        assert_true( size_t(ptr-flu) <= cnt );
         gym::unmapBufferC4VD(0);
         gym::drawPoints(pointSize, 0, ptr-flu);
     }
@@ -220,7 +220,7 @@ void Display1::drawSinglesA(const SingleSet & set) const
             ptr += 2;
         }
     }
-    assert_true( ptr-flu <= cnt );
+    assert_true( size_t(ptr-flu) <= cnt );
     gym::unmapBufferC4VD(0);
 
     if ( prop->link_width > 0 )
@@ -271,7 +271,7 @@ void Display1::drawCouplesF1(CoupleSet const& set) const
             shiftVertex(f4d, obj);
         }
     }
-    assert_true( ptr-flu <= cnt );
+    assert_true( size_t(ptr-flu) <= cnt );
     gym::unmapBufferC4VD(0);
     gym::drawPoints(pointSize, 0, ptr-flu);
     // display inactive Couples with square dots:
@@ -322,7 +322,7 @@ void Display1::drawCouplesF2(CoupleSet const& set) const
             ++ptr;
         }
     }
-    assert_true( ptr-flu <= cnt );
+    assert_true( size_t(ptr-flu) <= cnt );
     gym::unmapBufferC4VD(0);
     gym::drawPoints(pointSize, 0, ptr-flu);
     gym::cleanup();
@@ -357,7 +357,7 @@ void Display1::drawCouplesA(CoupleSet const& set) const
                 ++ptr;
             }
         }
-        assert_true( ptr-flu <= cnt );
+        assert_true( size_t(ptr-flu) <= cnt );
         gym::unmapBufferC4VD(0);
         gym::drawPoints(pointSize, 0, ptr-flu);
         gym::cleanup();
@@ -408,7 +408,7 @@ void Display1::drawCouplesB1(CoupleSet const& set) const
         }
 #endif
     }
-    assert_true( ptr-flu <= cnt );
+    assert_true( size_t(ptr-flu) <= cnt );
     gym::unmapBufferC4VD(0);
     
     if ( prop->link_width > 0 )
@@ -444,7 +444,7 @@ void Display1::drawCouplesB0(CoupleSet const& set) const
         ptr[1] = { obj->disp2()->color, Q };
         ptr += 2;
     }
-    assert_true( ptr-flu <= cnt );
+    assert_true( size_t(ptr-flu) <= cnt );
     gym::unmapBufferC4VD(1);
     if ( prop->link_width > 0 )
         gym::drawLines(linkWidth, 0, ptr-flu);
